requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
email_validator==2.1.0.post1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
WTForms==3.1.1
gunicorn==21.2.0
